Basic support for gopherjs/wasm and webgl backend.
To try, you can use opherjs build for one of the command, but the nicest
way is with `gopherjs serve fyne.io/fyne`. That last command will start
a local web server that you can use to navigate fyne directory and check
any app with it. Otherwise you can build with `GOOS=js GOARCH=wasm go
build` for the WebAssembly backend.
