+++ app/settings.go
@@ -2,8 +2,6 @@ package app
 
 import (
 	"bytes"
-	"encoding/json"
-	"io"
 	"os"
 	"path/filepath"
 	"sync"
@@ -106,28 +104,6 @@ func (s *settings) apply() {
 		}
 		return true
 	})
-}
-
-func (s *settings) load() {
-	err := s.loadFromFile(s.schema.StoragePath())
-	if err != nil && err != io.EOF { // we can get an EOF in windows settings writes
-		fyne.LogError("Settings load error:", err)
-	}
-
-	s.setupTheme()
-}
-
-func (s *settings) loadFromFile(path string) error {
-	file, err := os.Open(path) // #nosec
-	if err != nil {
-		if os.IsNotExist(err) {
-			return nil
-		}
-		return err
-	}
-	decode := json.NewDecoder(file)
-
-	return decode.Decode(&s.schema)
 }
 
 func (s *settings) fileChanged() {
