+++ internal/driver/glfw/clipboard.go
@@ -1,3 +1,6 @@
+//go:build !js && !wasm && !web
+// +build !js,!wasm,!web
+
 package glfw
 
 import (
