+++ internal/driver/glfw/glfw_es.go
@@ -1,6 +1,9 @@
-//go:build (gles || arm || arm64) && !darwin
+//go:build (gles || arm || arm64) && !darwin && !js && !wasm && !web
 // +build gles arm arm64
 // +build !darwin
+// +build !js
+// +build !wasm
+// +build !web
 
 package glfw
